Repair SwitchMode 6
Now "SwitchMode 6" works so that on every first click after restarting the timer it returns HOLD no matter if we hold the button or just press. It seems to me a bug but if it was intended, please add my code as a new "SwitchMode". Sorry for my English (I'm helping myself Google Translator)